+++ src/internal/transactionSubmission.ts
@@ -10,9 +10,9 @@ import { MoveVector, U8 } from "../bcs";
 import { postAptosFullNode } from "../client";
 import { Account } from "../core/account";
 import { AccountAddress, AccountAddressInput } from "../core/accountAddress";
-import { PrivateKey, Signature } from "../core/crypto";
+import { PrivateKey } from "../core/crypto";
 import { AccountAuthenticator } from "../transactions/authenticator/account";
-import { ProofChallenge, RotationProofChallenge } from "../transactions/instances/rotationProofChallenge";
+import { RotationProofChallenge } from "../transactions/instances/rotationProofChallenge";
 import {
   buildTransaction,
   generateTransactionPayload,
@@ -208,13 +208,6 @@ export function signTransaction(args: { signer: Account; transaction: AnyRawTran
   return accountAuthenticator;
 }
 
-export function signProofChallenge(args: { challenge: ProofChallenge; signer: Account }): Signature {
-  const { challenge, signer } = args;
-  const challengeHex = challenge.bcsToBytes();
-  const signature = signer.sign(challengeHex);
-  return signature;
-}
-
 /**
  * Simulates a transaction before singing it.
  *
@@ -357,7 +350,6 @@ export async function rotateAuthKey(args: {
 
   // Sign the challenge
   const challengeHex = challenge.bcsToBytes();
-
   const proofSignedByCurrentPrivateKey = fromAccount.sign(challengeHex);
   const proofSignedByNewPrivateKey = newAccount.sign(challengeHex);
