+++ tests/e2e/transaction/signTransaction.test.ts
@@ -6,8 +6,6 @@ import {
   AccountAuthenticator,
   AccountAuthenticatorEd25519,
   AccountAuthenticatorSingleKey,
-  MoveVector,
-  U8,
 } from "../../../src";
 import { longTestTimeout } from "../../unit/helper";
 import { getAptosClient } from "../helper";
@@ -217,47 +215,4 @@ describe("sign transaction", () => {
       });
     });
   });
-});
-
-test.only("test", async () => {
-  const fromAccount = Account.generate();
-  const newAccount = Account.generate();
-
-  await aptos.fundAccount({ accountAddress: fromAccount.accountAddress, amount: 1_000_000_000 });
-  await aptos.fundAccount({ accountAddress: newAccount.accountAddress, amount: 1_000_000_000 });
-
-  const accountInfo = await aptos.getAccountInfo({
-    accountAddress: fromAccount.accountAddress,
-  });
-
-  const challenge = await aptos.createProofChallenge({
-    struct: "0x1::account::RotationProofChallenge",
-    data: [
-      BigInt(accountInfo.sequence_number),
-      fromAccount.accountAddress,
-      accountInfo.authentication_key,
-      newAccount.publicKey.toUint8Array(),
-    ],
-  });
-
-  const proofSignedByCurrentPrivateKey = aptos.signProofChallenge({ challenge, signer: fromAccount });
-  const proofSignedByNewPrivateKey = aptos.signProofChallenge({ challenge, signer: newAccount });
-
-  const transaction = await aptos.transaction.build.simple({
-    sender: fromAccount.accountAddress,
-    data: {
-      function: "0x1::account::rotate_authentication_key",
-      functionArguments: [
-        new U8(fromAccount.signingScheme), // from scheme
-        MoveVector.U8(fromAccount.publicKey.toUint8Array()),
-        new U8(newAccount.signingScheme), // to scheme
-        MoveVector.U8(newAccount.publicKey.toUint8Array()),
-        MoveVector.U8(proofSignedByCurrentPrivateKey.toUint8Array()),
-        MoveVector.U8(proofSignedByNewPrivateKey.toUint8Array()),
-      ],
-    },
-  });
-
-  const response = await aptos.signAndSubmitTransaction({ signer: fromAccount, transaction });
-  console.log("response", response);
 });
